+++ include/hs_routes.h
@@ -6,6 +6,7 @@
 #include "hs_routes_common.h"
 #include "hs_routes_error.h"
 #include "hs_routes_fs.h"
+#include "hs_routes_payload_limit.h"
 #include "hs_routes_powered_by.h"
 #include "hs_routes_session.h"