+++ tests/test_server_serve.c
@@ -89,6 +89,9 @@ void test_impl()
     server->request_recv_timeout_seconds = 5;
     struct sockaddr_in address = hs_server_init_ipv4_address(7005);
 
+    hs_router_add_route(server->router, hs_routes_error_411_length_required_route_new());
+    hs_router_add_route(server->router, hs_routes_payload_limit_route_new(1024));
+
     route->is_get = true;
     route->path   = strdup("/gohome");
     route->serve  = _test_redirect;
@@ -104,7 +107,7 @@ void test_impl()
     route->path = strdup("/admin/");
 
     hs_router_add_route(server->router, route);
-    route = hs_routes_error_404_route_new();
+    route = hs_routes_error_404_not_found_route_new();
     hs_router_add_route(server->router, route);
 
     bool done = hs_server_serve(server, address, "test context", _test_should_stop_server, _test_should_stop_router);