+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 * New route flow state which can be used to store data as part of the request handling flow
 * Renaming common routes to enable internal refactoring
 * Close the socket if payload exists but not read after response is sent.
+* New 411 route.
+* New payload limit route.
 
 ### v0.2.3 (2021-08-24)